Generic Unit Tests for Objects
Created the most generic Unit Tests for every object available.

Skipped unit tests for images.

I aslo skipped unit tests for objects calling for another object. This is because the unit tests for the that object should already have unit tests for their parameters. Unit Testing an object call will require the object call, so I am assuming that it won't be a unit test anymore if it requires another object. Maybe?